Base
Created a simple base for the game by creating a canvas. I also
implemented a simple draw system that saves objects with draw data and
draws them. I will no doubt have to work with this system in the future,
especially when I later implement moving draw objects.

I also implemented a simple playlayout consisting of a playfield and a
menu field.