Entities
Removed the old drawObject arrays and replaced them with an entity
system. This will help in the future to track things like collision
between sprites. With this I also added the first version of the players
character. For now it's an immobile white square, but I have high hopes
for the future :)

Obviously not finished yet though.